src/test/java/pl/strategy/SmartStrategyTest.java
package pl.strategy;

import org.junit.jupiter.api.Test;
import pl.command.StringDrink;
import pl.command.StringRecipe;
import pl.observer.HumanClient;
import pl.observer.StringBar;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class SmartStrategyTest extends OrderTest{

    @Test
    public void smartStrategyStartOpened(){

        StringBar stringBar = new StringBar();
        StringDrink drink = new StringDrink("AbCd-aBcD");
        StringRecipe recipe = getRecipe();

        SmartStrategy strategy = new SmartStrategy();
        HumanClient client = new HumanClient(strategy);

        //Recipe is ordered immediately as happy hour was already under way
        stringBar.startHappyHour();
        client.wants(drink, recipe, stringBar);
        assertEquals("dCbX-DcBa", drink.getText());
    }

    @Test
    public void smartStrategyStartClosed(){

        StringBar stringBar = new StringBar();
        StringDrink drink = new StringDrink("AbCd-aBcD");
        StringRecipe recipe = getRecipe();

        SmartStrategy strategy = new SmartStrategy();
        HumanClient client = new HumanClient(strategy);
        stringBar.addObserver(client);

        client.wants(drink, recipe, stringBar);
        assertEquals("AbCd-aBcD", drink.getText());

        //Recipe is ordered here
        stringBar.startHappyHour();
        assertEquals("dCbX-DcBa", drink.getText());
    }
}
